+++ packages/terra-data-grid/src/index.js
@@ -1,0 +1,4 @@
+import WorklistDataGrid from './WorklistDataGrid';
+
+// eslint-disable-next-line import/prefer-default-export
+export { WorklistDataGrid };