+++ packages/terra-file-path/tests/wdio/file-path-spec.js
@@ -1,0 +1,11 @@
+Terra.describeViewports('FilePath', ['medium', 'large'], () => {
+  describe('placeholder', () => {
+    beforeEach(() => {
+      browser.url('/raw/tests/cerner-terra-framework-docs/file-path/default-file-path');
+    });
+
+    it('placeholder', () => {
+      Terra.validates.element('placeholder');
+    });
+  });
+});