+++ packages/terra-framework-docs/src/terra-dev-site/doc/tabs/example/TabsTemplate.jsx
@@ -58,10 +58,9 @@ const TabsTemplate = (props) => {
   const disabledTab = (
     <Tabs.Pane label="Disabled Tab" icon={<IconSearch />} isDisabled key="DisabledTab" />
   );
-
   return (
     <div className={cx('content-wrapper')}>
-      <Tabs defaultActiveKey="LabelTab" id={props.id}>
+      <Tabs setFocusOnContent defaultActiveKey="LabelTab" id={props.id}>
         {labelTab}
         {iconTab}
         {customTab}