+++ packages/terra-tabs/tests/wdio/tabs-spec.js
@@ -166,6 +166,28 @@ Terra.describeViewports('Tabs - Drag and Drop', ['medium'], () => {
   });
 });
 
+Terra.describeViewports('Tabs - Add and Close', ['medium'], () => {
+  it('display tabs with add icon', () => {
+    browser.url('/raw/tests/cerner-terra-framework-docs/tabs/tabs/add-close-tab');
+    Terra.validates.element('Tabs with Add', { selector: '#root' });
+  });
+
+  it('After Adding New Tabs', () => {
+    $('[data-terra-tab-add-btn="true"]').click();
+    Terra.validates.element('New Tab Added', { selector: '#root' });
+  });
+
+  it('display tabs with add and close icon', () => {
+    browser.url('/raw/tests/cerner-terra-framework-docs/tabs/tabs/add-close-tab');
+    Terra.validates.element('Tabs with close', { selector: '#root' });
+  });
+
+  it('display closable tabs', () => {
+    browser.keys(['Tab', 'ArrowRight', 'ArrowRight', 'ArrowRight', 'Delete']);
+    Terra.validates.element('Tab after close', { selector: '#root' });
+  });
+});
+
 Terra.describeViewports('Tabs - Drag and Drop', ['tiny'], () => {
   it('displays tab before drag and drop', () => {
     browser.url('/raw/tests/cerner-terra-framework-docs/tabs/tabs/controlled-tabs');
@@ -186,5 +208,14 @@ Terra.describeViewports('Tabs - Programmatic Activation', ['medium'], () => {
   it('should perform drag and drop operation', () => {
     $('#activate-tab').click();
     Terra.validates.element('After Programmatic Activation', { selector: '#root' });
+  });
+});
+
+Terra.describeViewports('Tabs - Keyboard Focus with interactive element', ['medium'], () => {
+  it('displays tab with keyboard focus with interactive element within the tab content', () => {
+    browser.url('/raw/tests/cerner-terra-framework-docs/tabs/tabs/add-close-interactive-tab');
+    browser.keys('Tab');
+    browser.keys('Tab');
+    Terra.validates.element('Keyboard Focus on First Interactive Element', { selector: '#root' });
   });
 });