+++ packages/terra-date-picker/CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed DatePicker scroll on smaller viewports.
+
+* Updated
+  * Screen reader announcement for visual label.
+
+## 4.89.0 - (September 5, 2023)
+
+* Changed
+  * Visual focus changed to dashed lines for individual fields.
+  * Fixed the date format being read 3 times when using SR
+
+## 4.88.0 - (August 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 4.87.1 - (August 4, 2023)
 
 * Changed