+++ packages/terra-dialog-modal/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 ## Unreleased
 
+## 3.91.0 - (September 5, 2023)
+
+* Changed
+  * Minor dependency version bump
+
+## 3.90.0 - (August 31, 2023)
+
+* Changed
+  * Minor dependency version bump
+
 ## 3.89.1 - (August 4, 2023)
 
 * Changed